# Conflicts:
#	resources/js/components/Course/CourseViewerComponent.vue
#	resources/js/components/Course/SpeakingComponent.vue

+++ database/migrations/0006_00_00_000000_create_lecture_table.php
@@ -15,7 +15,7 @@ class CreateLectureTable extends Migration
     {
         Schema::create('lectures', function (Blueprint $table) {
             $table->id();
-            $table->integer('level')->nullable(true);
+            $table->double('level')->nullable(true);
             $table->integer('type')->nullable(true);
             $table->integer('index')->default(0);
             $table
